Break solver tests into separate tests
